@@ -7,20 +7,25 @@
         {
             "name": "Esron Dtamar",
             "email": "esron.dtamar@gmail.com"
+        },
+        {
+            "name": "Gustavo Lemes Saraiva",
+            "email" : "gustavolemes2@gmail.com"
         }
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=7.1",
-        "monolog/monolog": "^1.12",
-        "guzzlehttp/guzzle": "^6.0"
+        "php": "^8.0",
+        "monolog/monolog": "^2.0",
+        "guzzlehttp/guzzle": "^7.0",
+        "illuminate/support": "^8.40"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.3"
     },
     "autoload": {
         "psr-4": {
-            "Sysvale\\Logging\\": "src/"
+            "Drugovich\\Logging\\": "src/"
         }
     },
     "autoload-dev": {
